% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.R
\name{as.kpilist}
\alias{as.kpilist}
\title{Convert a list to a kpilist}
\usage{
as.kpilist(x)
}
\arguments{
\item{x}{list of kpi objects}
}
\value{
a kpilist
}
\description{
Convert a list to a kpilist
}
\examples{
l <- lapply(c("drat", "hp", "qsec"), function(x){
 kpi(mtcars,
     var = x,
     by = c("am", "cyl"),
     kpi_fn = kpi_fn_median)
})
as.kpilist(l)

}
