% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_kpi.R
\name{calc_kpi}
\alias{calc_kpi}
\title{Calculate KPIs}
\usage{
calc_kpi(data, var, by = NULL, kpi_fn, txt = "")
}
\arguments{
\item{data}{dataframe containing \code{var} and, optionally, \code{by}}

\item{var}{string, variable to calculate KPI from}

\item{by}{a stratifying variable, KPIs calculated for each strata seperately}

\item{kpi_fn}{summary function}

\item{txt}{textual description of the KPI}
}
\value{
a data frame of class \code{kpicalc}
}
\description{
Calculate KPIs
}
\examples{
# data(mtcars)
# mtcars \%>\% kpitools::calc_kpi("mpg", by = "am", kpi_fn = kpi_fn_median)
}
