% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fab_dow.R
\name{fab_dow}
\alias{fab_dow}
\title{Day of week figure(s)}
\usage{
fab_dow(
  data,
  var,
  by = NULL,
  dow_fmt = "\%a",
  output = c("facet", "list"),
  col = "grey",
  fill = "grey",
  ...
)
}
\arguments{
\item{data}{cdata frame containing \code{var} (and, optionally, \code{by}) variable(s)}

\item{var}{string. Name of variable containing relevant dates or datetimes
(will be coerced to date via \code{as.Date})}

\item{by}{string. Name of variable denoting grouping}

\item{dow_fmt}{format for day of week}

\item{output}{output format \code{facet} combines figures via ggplot2::facet_wrap,
\code{list} returns a list of ggplot2 plots}

\item{col}{colour to use for bar lines}

\item{fill}{colour to use for bar fill}

\item{...}{options passed to facet_wrap (see examples)}
}
\value{
list or ggplot2 object
}
\description{
In a normal setting it may be that observations that occur at the weekend are
indicative of data fabrication. \code{fab_dow} (short for fabrication, day of
week), produces a plot that may help to identify problems. Customs vary in
different countries, so that should be accounted for when interpreting these
figures.
}
\examples{
set.seed(234)
dat <- data.frame(
  x = Sys.Date() + sample(-20:19, 40, TRUE),
  by = c(rep(1, 10), rep(2, 30))
)
dat \%>\% fab_dow("x")
dat \%>\% fab_dow("x", "by")
# free x scale
dat \%>\% fab_dow("x", "by", scales = "free_x")
# different colour bars
dat \%>\% fab_dow("x", fill = "orange")
# list of plots
dat \%>\% fab_dow("x", "by", output = "list")
# change colours
dat \%>\% fab_dow("x", col = "purple", fill = "pink")

}
