% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fab_tod.R
\name{fab_tod}
\alias{fab_tod}
\title{Time of day figure(s)}
\usage{
fab_tod(
  data,
  var,
  by = NULL,
  dow_fmt = "\%a",
  output = c("list", "facet"),
  col_poly = "black",
  x_poly = c(8.5, 21.5),
  col_bars = "grey"
)
}
\arguments{
\item{data}{data frame containing \code{var} (and, optionally, \code{by}) variable(s)}

\item{var}{string. Name of variable containing relevant datetimes}

\item{by}{string. Name of variable denoting grouping}

\item{dow_fmt}{format for day of week}

\item{output}{output format \code{facet} combines figures via ggplot2::facet_wrap,
\code{list} returns a list of ggplot2 plots}

\item{col_poly}{colour to use for the region indicating possible fabrication}

\item{x_poly}{x coordinates for the start and end of the region indicating possible fabrication}

\item{col_bars}{colour to use for bars indicating counts}
}
\value{
list or ggplot2 object
}
\description{
In a normal setting it may be that observations that occur at night are
indicative of data fabrication. \code{fab_tod} (short for fabrication, time
of day), produces a plot that may help to identify problems. Customs vary in
different countries, so that should be accounted for when interpreting these
figures.
}
\details{
Due to a limitation of faceting plots with polar coordinates, faceted
plots all have the same y coordinate (equivalent to fixed axes). To free the
coordinate system, use the list output (default) and wrap them together using
e.g. patchwork, possibly applying some customizations in advance.
}
\examples{
set.seed(234)
dat <- data.frame(
  x = lubridate::ymd_h("2020-05-01 14") + 60^2*sample(0:20, 40, TRUE),
  by = c(rep(1, 10), rep(2, 30))
)
dat \%>\% fab_tod("x")
dat \%>\% fab_tod("x") + theme_kpitools()
dat \%>\% fab_tod("x", "by")
#faceted of plots
dat \%>\% fab_tod("x", "by", output = "facet")
#with patchwork
patchwork::wrap_plots(dat \%>\% fab_tod("x", "by"))

}
