% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{kpi}
\alias{kpi}
\title{Create KPI tables}
\usage{
kpi(
  data,
  var,
  by = NULL,
  kpi_fn = kpi_fn_mean,
  txt = "",
  cutpoints = NULL,
  cutlabels = NULL
)
}
\arguments{
\item{cutlabels}{}
}
\value{

}
\description{
Create KPI tables
}
