% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kpi_fn_summary_fns.R
\name{kpi_fn_n}
\alias{kpi_fn_n}
\alias{Proportions}
\alias{kpi_fn_prop}
\alias{Percentages}
\alias{kpi_fn_perc}
\alias{Median}
\alias{kpi_fn_median}
\alias{Mean}
\alias{kpi_fn_mean}
\alias{IQR}
\alias{kpi_fn_iqr}
\alias{Minimum}
\alias{kpi_fn_min}
\alias{Maximum}
\alias{kpi_fn_max}
\title{KPI summary functions}
\usage{
kpi_fn_n(.data)

kpi_fn_prop(.data)

kpi_fn_perc(.data)

kpi_fn_median(.data)

kpi_fn_mean(.data)

kpi_fn_iqr(.data)

kpi_fn_min(.data)

kpi_fn_max(.data)
}
\arguments{
\item{.data}{data frame}
}
\value{

}
\description{
KPI summary functions
}
