% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kpi_fn_summary_fns.R
\name{kpi_fns}
\alias{kpi_fns}
\title{Get a list of KPI summary functions provided by \code{kpitools}.}
\usage{
kpi_fns()
}
\value{
character vector of functions
}
\description{
Get a list of KPI summary functions provided by \code{kpitools}.
}
\examples{
kpi_fns()
}
\seealso{
kpi_fn_
}
