% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kpi_gauge.R
\name{kpi_gauge}
\alias{kpi_gauge}
\title{Depict KPIs as gauges}
\usage{
kpi_gauge(
  kpi,
  level = "overall",
  focus = NULL,
  annot_size = 8,
  title = TRUE,
  others = FALSE,
  others_alpha = 0.25,
  needle_width = 1,
  cont_colours = c("green", "gold", "red")
)
}
\arguments{
\item{kpi}{\code{kpi} object}

\item{level}{which level to depict (e.g. \code{overall}, or one of the \code{by} variables)}

\item{focus}{if using a \code{by} variable, which 'site' to make the focus}

\item{others}{logical, depict the other 'sites' (levels of the \code{by} variable)}

\item{others_alpha}{alpha to use for the \code{others}}

\item{needle_width}{Width of the needle}

\item{cont_colours}{for continuous scales, a set of colours to use for the ramp}
}
\value{
\code{ggplot2} object
}
\description{
THIS IS EXPERIMENTAL - expect changes/bugs
}
\examples{
mtcars \%>\%
mutate(cylgt4 = cyl > 4) \%>\%
  kpi(var = "mpg",
      breakpoints = c(0, 22, 50),
      by = c("am", "cyl"),
      txt = "MPG",
      kpi_fn = kpi_fn_median) \%>\%
  kpi_gauge("am", 0,
            needle_width = .2)
mtcars \%>\%
  mutate(cylgt4 = cyl > 4) \%>\%
  kpi(var = "mpg",
      by = c("am", "cyl"),
      txt = "MPG",
      kpi_fn = kpi_fn_median
      #, dir = "dec"
      ) \%>\%
  kpi_gauge("cyl", 6,
            others = TRUE,
            needle_width = .5)
}
