% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kpi_outlier.R
\name{kpi_outlier}
\alias{kpi_outlier}
\title{Get the outliers}
\usage{
kpi_outlier(kpitab, n_iqr = 2)
}
\arguments{
\item{kpitab}{result from calc_kpi}

\item{n_iqr}{number of IQRs below/above the lower/upper quartiles that should be considered outliers}
}
\value{
\code{kpitab} without the outliers
}
\description{
Get the outliers
}
\examples{
#' data(mtcars)
mtcars \%>\%
  calc_kpi("mpg", by = "am", kpi_fn = kpi_fn_median) \%>\%
  kpi_outlier()
}
