% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.kpi}
\alias{plot.kpi}
\title{Plot KPI objects}
\usage{
\method{plot}{kpi}(x, y = 1, col = "#E6002EFF", pch = 21, ...)
}
\arguments{
\item{x}{result from kpi}

\item{y}{ignored}

\item{col}{colour for points}

\item{pch}{point character}

\item{...}{for possible future expansion}
}
\value{
list of ggplot objects
}
\description{
Plot KPI objects
}
\examples{
# defaults
kpi <- mtcars \%>\%
  kpi("mpg", by = c("am", "vs"), txt = "MPG")
plot(kpi)

# customizing the plots
plots <- plot(kpi)

plots$am +
  theme_bw() +
  labs(title = "Foo")

}
