% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.kpi}
\alias{print.kpi}
\title{Print method for kpi objects}
\usage{
\method{print}{kpi}(x, table = TRUE, outlier = TRUE, ...)
}
\arguments{
\item{x}{kpi object}

\item{table}{logical, whether to add a table stats by grouping variable(s) to the output}

\item{outlier}{logical, whether to add a table of outliers to the output}

\item{...}{not currently used}
}
\value{
output printed to the console
}
\description{
Print method for kpi objects
}
