% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kpi_cut.R
\name{riskcols}
\alias{riskcols}
\title{Colors for KPIs cutoffs}
\usage{
riskcols(x)
}
\arguments{
\item{x}{breakpoints}
}
\value{
string of \code{length(x) - 1} with suitable colors.
}
\description{
Colors for KPIs cutoffs
}
\examples{
riskcols(1:4)
}
