% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kpi_cut.R
\name{risklabs}
\alias{risklabs}
\title{Labels for KPIs with cutoffs}
\usage{
risklabs(x)
}
\arguments{
\item{x}{breakpoints}
}
\value{
string of \code{length(x) - 1} with suitable labels.
}
\description{
Labels for KPIs with cutoffs
}
\examples{
risklabs(1:4)
}
