% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_kpitools.R
\name{theme_kpitools}
\alias{theme_kpitools}
\title{kpitools \code{ggplot2} theme}
\usage{
theme_kpitools()
}
\value{
\code{ggplot2} theme object
}
\description{
Theme based on \code{theme_bw} and removing y-axis and moving the legend to beneath the plot.
}
\examples{
kpi <- mtcars \%>\%
  kpi("mpg", by = "cyl", txt = "MPG")

# without the theme
plot(kpi)$cyl
# with the theme
plot(kpi)$cyl +
  theme_kpitools()
}
