% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_data_frames.R
\name{as_data_frames}
\alias{as_data_frames}
\title{Convert a kpilist to a set of dataframes for overall numbers and each \code{by} group}
\usage{
as_data_frames(x, keep_kpiname = TRUE, ...)
}
\arguments{
\item{x}{kpilist}

\item{keep_kpiname}{each kpi in a \code{kpilist} has a name constructed from
the variable and the summary function. Should this name be retained in the
output?}

\item{...}{currently ignored}
}
\value{
list of dataframes
}
\description{
Convert a kpilist to a set of dataframes for overall numbers and each \code{by} group
}
\examples{
kpi1 <- mtcars \%>\%
  kpi(var = "mpg", by = c("am", "cyl"), txt = "MPG",
      kpi_fn = kpi_fn_median)
kpi2 <- mtcars \%>\%
  kpi(var = "drat", by = c("am", "cyl"), txt = "DRAT",
      kpi_fn = kpi_fn_median)
kpi3 <- mtcars \%>\%
  mutate(cylgt4 = cyl > 4) \%>\%
  kpi(var = "cylgt4", by = c("am", "cyl"), txt = "Cylinders",
      kpi_fn = kpi_fn_perc)
l <- c(kpi1, kpi2, kpi3)
kpitools:::as_data_frames(l, FALSE)

}
