% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{kpi}
\alias{kpi}
\title{Create KPI tables}
\usage{
kpi(
  data,
  var,
  by = NULL,
  kpi_fn = kpi_fn_mean,
  txt = "",
  n_iqr = 2,
  breakpoints = NULL,
  risklabels = risklabs(breakpoints),
  direction = c("increasing", "decreasing"),
  raw_cut = FALSE
)
}
\arguments{
\item{data}{a data frame}

\item{var}{the variable to summarize}

\item{by}{optional variable(s) to group over}

\item{kpi_fn}{summary function}

\item{txt}{a descriptive text}

\item{n_iqr}{number of IQRs below/above the lower/upper quartiles that should be considered outliers}

\item{breakpoints}{cut points (if KPIs use a traffic light system)}

\item{risklabels}{labels for the cut points. By default, variations on
low/moderate/high are used}

\item{direction}{seriousness relative to \code{breakpoints}}

\item{raw_cut}{add a group variable without applying \code{risklabels}}
}
\value{
a list with either 1 or (length(by) + 1) lists.
}
\description{
Create KPI tables
}
\examples{
kpi <- mtcars \%>\%
  mutate(cylgt4 = cyl > 4) \%>\%
  kpi(var = "mpg",
      breakpoints = c(0, 22, 50),
      by = c("am", "cyl"),
      txt = "MPG",
      kpi_fn = kpi_fn_median)
}
