% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kpi_cut.R
\name{kpi_cut}
\alias{kpi_cut}
\title{Cut KPI indicators}
\usage{
kpi_cut(
  kpitab,
  breakpoints,
  risklabels = risklabs(breakpoints),
  direction = c("increasing", "decreasing"),
  raw_cut = FALSE
)
}
\arguments{
\item{kpitab}{output from calc_kpi}

\item{breakpoints}{cut points (if KPIs use a traffic light system)}

\item{risklabels}{labels for the cut points. By default, variations on
low/moderate/high are used}

\item{direction}{seriousness relative to \code{breakpoints}}

\item{raw_cut}{add a group variable without applying \code{risklabels}}
}
\value{

}
\description{
Cut KPI indicators
}
\examples{
kpitab <- mtcars \%>\%
   kpitools:::kpi_calc("mpg", by = "am",
     kpi_fn = kpi_fn_median, txt = "MPG")
cutpoints <- c(0, 20, 30)
kpitools:::kpi_cut(kpitab, cutpoints, risklabels = c("Low", "High"))
kpitools:::kpi_cut(kpitab, cutpoints, risklabels = 1:2)
kpitools:::kpi_cut(kpitab, cutpoints)
}
\keyword{internal}
