% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kpi_cut.R
\name{kpi_cut}
\alias{kpi_cut}
\title{Cut KPI indicators}
\usage{
kpi_cut(kpitab, cutpoints, cutlabels = NULL)
}
\arguments{
\item{kpitab}{output from calc_kpi}

\item{cutpoints}{cut points (if KPIs use a traffic light system)}

\item{cutlabels}{labels for the cut points}
}
\value{

}
\description{
Cut KPI indicators
}
\examples{
# kpitab <- mtcars \%>\% calc_kpi("mpg", by = "am", kpi_fn = kpi_fn_median, txt = "MPG")
# cutpoints <- c(0, 20, 30)
# kpi_cut(kpitab, cutpoints, cutlabels = c("Low", "High"))
# kpi_cut(kpitab, cutpoints, cutlabels = 1:2)
# kpi_cut(kpitab, cutpoints)
}
