% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kpi_fn_summary_fns.R
\name{kpi_fn_n}
\alias{kpi_fn_n}
\alias{Proportions}
\alias{kpi_fn_prop}
\alias{Percentages}
\alias{kpi_fn_perc}
\alias{Median}
\alias{kpi_fn_median}
\alias{Mean}
\alias{kpi_fn_mean}
\alias{IQR}
\alias{kpi_fn_iqr}
\alias{Minimum}
\alias{kpi_fn_min}
\alias{Maximum}
\alias{kpi_fn_max}
\title{KPI summary functions}
\usage{
kpi_fn_n(.data)

kpi_fn_prop(.data)

kpi_fn_perc(.data)

kpi_fn_median(.data)

kpi_fn_mean(.data)

kpi_fn_iqr(.data)

kpi_fn_min(.data)

kpi_fn_max(.data)
}
\arguments{
\item{.data}{data frame}
}
\value{

}
\description{
These functions are not intended to be run as they are. They are intended to
be passed as arguments to the \code{kpi} or \code{calc_kpi} functions. They
summarize the data in the appropriate manner for the type of KPI. For example,
the \code{kpi_fn_prop} counts the number of cases and total number of observations
then calculates a proportion. \code{kpi_fn_median} simply calculates the median
of the observations.
}
\details{
Functions should accept a dataframe with a \code{var} variable and return a
dataframe with \code{stat} (other variables are optional, although an \code{N}
variable allows for compatibility with downstream functions). All provided
functions return \code{stat} and \code{N}, with some also returning \code{n}.
}
